[2/n] tls: refactor tlsscenario under tls pkg (#2162)
